Allow omniauth-oauth2 1.2.0
This version was released on July 9, 2014 and is required for several
other providers. I went with a relaxing of the requirement instead of
just a bump in requirement in case some people are locked to 1.1.1 by
another provider. This seems to be the case when using many omniauth
providers -- a conflict in the required versions in other omniauth
libraries.